Fix conjure packages (#105)
