--- extras/jackson-support/src/test/java/com/palantir/remoting/api/ext/jackson/ShimJdk7ModuleTest.java
@@ -1,35 +1,0 @@
-/*
- * (c) Copyright 2017 Palantir Technologies Inc. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.palantir.remoting.api.ext.jackson;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import java.io.IOException;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import org.junit.Test;
-
-public final class ShimJdk7ModuleTest {
-
-    @Test
-    public void testPathSerDe() throws IOException {
-        ObjectMapper mapper = new ObjectMapper().registerModule(new ShimJdk7Module());
-        String val = mapper.writeValueAsString(Paths.get("a", "b", "c"));
-        assertThat(mapper.readValue(val, Path.class)).isEqualTo(Paths.get("a", "b", "c"));
-    }
-}
