+++ service-config/src/test/java/com/palantir/conjure/java/api/config/service/ServiceConfigurationFactoryTests.java
@@ -27,8 +27,8 @@ import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.io.Resources;
 import com.palantir.conjure.java.api.config.ssl.SslConfiguration;
-import com.palantir.remoting.api.ext.jackson.ObjectMappers;
-import com.palantir.remoting.api.ext.jackson.ShimJdk7Module;
+import com.palantir.conjure.java.api.ext.jackson.ObjectMappers;
+import com.palantir.conjure.java.api.ext.jackson.ShimJdk7Module;
 import com.palantir.tokens.auth.BearerToken;
 import java.io.IOException;
 import java.net.URL;
