+++ ssl-config/src/test/java/com/palantir/conjure/java/api/config/ssl/SslConfigurationTest.java
@@ -19,7 +19,7 @@ package com.palantir.conjure.java.api.config.ssl;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.palantir.remoting.api.ext.jackson.ObjectMappers;
+import com.palantir.conjure.java.api.ext.jackson.ObjectMappers;
 import java.nio.file.Paths;
 import org.junit.Test;
